+++ src/test/java/org/jabref/logic/importer/fileformat/PdfMergeMetadataImporterTest.java
@@ -6,8 +6,8 @@ import java.util.Collections;
 import java.util.List;
 
 import org.jabref.logic.importer.ImportFormatPreferences;
+import org.jabref.logic.importer.importsettings.ImportSettingsPreferences;
 import org.jabref.logic.util.StandardFileType;
-import org.jabref.logic.xmp.XmpPreferences;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.StandardField;
 import org.jabref.model.entry.field.UnknownField;
@@ -25,13 +25,15 @@ class PdfMergeMetadataImporterTest {
 
     private PdfMergeMetadataImporter importer;
     private ImportFormatPreferences importFormatPreferences;
-    private XmpPreferences xmpPreferences;
 
     @BeforeEach
     void setUp() {
+        ImportSettingsPreferences importSettingsPreferences = mock(ImportSettingsPreferences.class, Answers.RETURNS_DEEP_STUBS);
+        when(importSettingsPreferences.isGrobidEnabled()).thenReturn(true);
+        when(importSettingsPreferences.getGrobidURL()).thenReturn("http://grobid.jabref.org:8070");
         importFormatPreferences = mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS);
         when(importFormatPreferences.getFieldContentFormatterPreferences().getNonWrappableFields()).thenReturn(List.of());
-        importer = new PdfMergeMetadataImporter(importFormatPreferences);
+        importer = new PdfMergeMetadataImporter(importSettingsPreferences, importFormatPreferences);
     }
 
     @Test