+++ src/test/java/org/jabref/logic/pdf/search/indexing/PdfIndexerTest.java
@@ -58,7 +58,7 @@ public class PdfIndexerTest {
 
         // then
         try (IndexReader reader = DirectoryReader.open(new NIOFSDirectory(context.getFulltextIndexPath()))) {
-            assertEquals(1, reader.numDocs());
+            assertEquals(33, reader.numDocs());
         }
     }
 
@@ -110,7 +110,7 @@ public class PdfIndexerTest {
 
         // then
         try (IndexReader reader = DirectoryReader.open(new NIOFSDirectory(context.getFulltextIndexPath()))) {
-            assertEquals(1, reader.numDocs());
+            assertEquals(33, reader.numDocs());
         }
     }
 
@@ -144,7 +144,7 @@ public class PdfIndexerTest {
         indexer.addToIndex(context);
         // index actually exists
         try (IndexReader reader = DirectoryReader.open(new NIOFSDirectory(context.getFulltextIndexPath()))) {
-            assertEquals(1, reader.numDocs());
+            assertEquals(33, reader.numDocs());
         }
 
         // when
@@ -168,7 +168,7 @@ public class PdfIndexerTest {
 
         // index with first entry
         try (IndexReader reader = DirectoryReader.open(new NIOFSDirectory(context.getFulltextIndexPath()))) {
-            assertEquals(1, reader.numDocs());
+            assertEquals(33, reader.numDocs());
         }
 
         BibEntry metadata = new BibEntry(StandardEntryType.Article);
@@ -180,7 +180,7 @@ public class PdfIndexerTest {
 
         // then
         try (IndexReader reader = DirectoryReader.open(new NIOFSDirectory(context.getFulltextIndexPath()))) {
-            assertEquals(2, reader.numDocs());
+            assertEquals(34, reader.numDocs());
         }
     }
 }